fix: 深度代码审核整改 - 安全加固、性能优化、Bug修复
安全修复:
- WebDAV 锁系统改用 Redis 分布式锁，支持多实例部署
- S3 签名比较改用 ConstantTimeCompare 防止时序攻击
- ApiKey Key 字段改为 json:"-" 防止 JSON 暴露
- S3 handler 统一使用 MapErrorToS3Error 类型安全映射
- S3 中间件日志截断 Authorization 头防止泄露
- 分享密码支持 POST body 传递，向后兼容 GET
- SSRF 防护增强，支持 IPv6 地址检查
- Content-Disposition 文件名过滤控制字符
- JWT token.Valid 检查、Cookie 域名精确匹配
- 密码修改后清除旧 session 使旧 token 失效
- session 清除失败记录 Warn 日志

Bug修复:
- DeleteDir 子目录删除失败时事务回滚，防止孤立目录
- download/file_package 双重写入修复(Seek 移到 Status 前)
- RenameFile TOCTOU 竞态修复，整体包裹事务
- CopyObject 数据库操作全部包裹事务
- AdminDeleteUser 级联清理所有关联数据
- recycled_service 物理删除使用 Unscoped()
- 中间状态解压任务 30 分钟超时自动清理
- State/Type 改为 *int 指针类型避免默认值 0 查询错误
- 磁盘空间检查改用实际可用空间而非配置大小

性能优化:
- file_package 打包下载 N+1 查询改为批量查询
- 新增 BatchGetByUserIDAndUfIDs 批量查询方法
- 所有 service 方法添加 30 秒 context 超时控制

前端修复:
- print.ts 使用 escapeHtml 防止 XSS
- permission.ts t 函数改为参数传入
- password.ts 改用 crypto.getRandomValues
- 管理员判断使用 ADMIN_GROUP_ID 常量
- uploadTaskManager 类型安全改进

依赖升级:
- Go 版本升级到 1.26.4
- 80+ 个依赖包升级到最新版本

+++ src/pkg/models/upload_chunk.go
@@ -5,7 +5,7 @@ type UploadChunk struct {
 	ChunkID  int    `json:"chunk_id" gorm:"primaryKey"`
 	UserID   string `json:"user_id"`
 	FileName string `json:"file_name"`
-	FileSize int    `json:"file_size"`
+	FileSize int64  `json:"file_size"`
 	Md5      string `json:"md5"`
 	PathID   string `json:"path_id"`
 }
